Move renderTaskRowHistogram out of index.
